# backend/serverPanels/swagger-output.json
{
    "swagger": "2.0",
    "info": {
        "title": "Nile Maret-Place Api",
        "description": "Description",
        "version": "1.0.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/rateVendorProducts": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "vendor_id": {
                                    "example": "any"
                                },
                                "customer_id": {
                                    "example": "any"
                                },
                                "product_uniqueid": {
                                    "example": "any"
                                },
                                "rating": {
                                    "example": "any"
                                },
                                "label": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/fetchRatings": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/addReview": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "example": "any"
                                },
                                "reviewText": {
                                    "example": "any"
                                },
                                "customer_id": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/emptyReviewMedia": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/uploadReviewMedia": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "rate_id": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    }
}